Refactored and added a bunch of tests
