+++ packages/flutter/test/cupertino/nav_bar_test.dart
@@ -306,6 +306,65 @@ void main() {
     expect(tester.getSize(find.widgetWithText(OverflowBox, 'Title')).height, 0.0);
   });
 
+  testWidgets('User specified middle is always visible in sliver', (WidgetTester tester) async {
+    final ScrollController scrollController = new ScrollController();
+    final Key segmentedControlsKey = new UniqueKey();
+    await tester.pumpWidget(
+      new CupertinoApp(
+        home: new CupertinoPageScaffold(
+          child: new CustomScrollView(
+            controller: scrollController,
+            slivers: <Widget>[
+              new CupertinoSliverNavigationBar(
+                middle: new ConstrainedBox(
+                  constraints: const BoxConstraints(maxWidth: 200.0),
+                  child: new SegmentedControl<int>(
+                    key: segmentedControlsKey,
+                    children: const <int, Widget>{
+                      0: const Center(child: const Text('Option A')),
+                      1: const Center(child: const Text('Option B')),
+                    },
+                    onValueChanged: (int selected) { },
+                    groupValue: 0,
+                  ),
+                ),
+                largeTitle: const Text('Title'),
+              ),
+              new SliverToBoxAdapter(
+                child: new Container(
+                  height: 1200.0,
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+
+    expect(scrollController.offset, 0.0);
+    expect(tester.getTopLeft(find.byType(NavigationToolbar)).dy, 0.0);
+    expect(tester.getSize(find.byType(NavigationToolbar)).height, 44.0);
+
+    expect(find.text('Title'), findsOneWidget);
+    expect(tester.getCenter(find.byKey(segmentedControlsKey)).dx, 400.0);
+
+    expect(tester.getTopLeft(find.widgetWithText(OverflowBox, 'Title')).dy, 44.0);
+    expect(tester.getSize(find.widgetWithText(OverflowBox, 'Title')).height, 52.0);
+
+    scrollController.jumpTo(600.0);
+    await tester.pump(); // Once to trigger the opacity animation.
+    await tester.pump(const Duration(milliseconds: 300));
+
+    expect(tester.getCenter(find.byKey(segmentedControlsKey)).dx, 400.0);
+    // The large title is invisible now.
+    expect(
+      tester.renderObject<RenderAnimatedOpacity>(
+        find.widgetWithText(AnimatedOpacity, 'Title')
+      ).opacity.value,
+      0.0,
+    );
+  });
+
   testWidgets('Small title can be overridden', (WidgetTester tester) async {
     final ScrollController scrollController = new ScrollController();
     await tester.pumpWidget(
