+++ packages/flutter/test/cupertino/route_test.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
@@ -34,6 +35,71 @@ void main() {
 
     // As a title, it should also be centered.
     expect(tester.getCenter(find.text('An iPod')).dx, 400.0);
+  });
+
+  testWidgets('Large title auto-populates with title', (WidgetTester tester) async {
+    await tester.pumpWidget(
+      new CupertinoApp(
+        home: const Placeholder(),
+      ),
+    );
+
+    tester.state<NavigatorState>(find.byType(Navigator)).push(
+      new CupertinoPageRoute<void>(
+        title: 'An iPod',
+        builder: (BuildContext context) {
+          return new CupertinoPageScaffold(
+            child: new CustomScrollView(
+              slivers: const <Widget>[
+                const CupertinoSliverNavigationBar(),
+              ],
+            ),
+          );
+        }
+      )
+    );
+
+    await tester.pump();
+    await tester.pump(const Duration(milliseconds: 500));
+
+    // There should be 2 Text widget with the title in the nav bar. One in the
+    // large title position and one in the middle position (though the middle
+    // position Text is initially invisible while the sliver is expanded).
+    expect(
+      find.widgetWithText(CupertinoSliverNavigationBar, 'An iPod'),
+      findsNWidgets(2),
+    );
+
+    final List<Element> titles = tester.elementList(find.text('An iPod'))
+        .toList()
+        ..sort((Element a, Element b) {
+          final RenderParagraph aParagraph = a.renderObject;
+          final RenderParagraph bParagraph = b.renderObject;
+          return aParagraph.text.style.fontSize.compareTo(
+            bParagraph.text.style.fontSize
+          );
+        });
+
+    final Iterable<double> opacities = titles.map((Element element) {
+      final RenderAnimatedOpacity renderOpacity =
+          element.ancestorRenderObjectOfType(const TypeMatcher<RenderAnimatedOpacity>());
+      return renderOpacity.opacity.value;
+    });
+
+    expect(opacities, <double> [
+        0.0, // Initially the smaller font title is invisible.
+        1.0, // The larger font title is visible.
+    ]);
+
+    // Check that the large font title is at the right spot.
+    expect(
+      tester.getTopLeft(find.byWidget(titles[1].widget)),
+      const Offset(16.0, 54.0),
+    );
+
+    // The smaller, initially invisible title, should still be positioned in the
+    // center.
+    expect(tester.getCenter(find.byWidget(titles[0].widget)).dx, 400.0);
   });
 
   testWidgets('Leading auto-populates with back button with previous title', (WidgetTester tester) async {
